Unify column sampling into one psi module
The merge of the client groundwork left two copies of the preview
sampling helpers; the server profile must not import from
components (the built server entry cannot resolve that alias), so
the superset module lands at psi/columnSamples and every consumer
imports it there.
